Split read-only tests into their own suite.
