+++ src/it/scala/com/databricks/spark/redshift/RedshiftWriteSuite.scala
@@ -1,0 +1,111 @@
+/*
+ * Copyright 2015 Databricks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.databricks.spark.redshift
+
+import java.sql.SQLException
+
+import org.apache.spark.sql._
+import org.apache.spark.sql.types._
+
+/**
+ * End-to-end tests of functionality which involves writing to Redshift via the connector.
+ */
+class RedshiftWriteSuite extends IntegrationSuiteBase {
+
+  test("roundtrip save and load") {
+    // This test can be simplified once #98 is fixed.
+    val tableName = s"roundtrip_save_and_load_$randomSuffix"
+    try {
+      write(
+        sqlContext.createDataFrame(sc.parallelize(TestUtils.expectedData), TestUtils.testSchema))
+        .option("dbtable", tableName)
+        .mode(SaveMode.ErrorIfExists)
+        .save()
+
+      assert(DefaultJDBCWrapper.tableExists(conn, tableName))
+      checkAnswer(read.option("dbtable", tableName).load(), TestUtils.expectedData)
+    } finally {
+      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
+      conn.commit()
+    }
+  }
+
+  test("roundtrip save and load with uppercase column names") {
+    testRoundtripSaveAndLoad(
+      s"roundtrip_write_and_read_with_uppercase_column_names_$randomSuffix",
+      sqlContext.createDataFrame(sc.parallelize(Seq(Row(1))),
+        StructType(StructField("A", IntegerType) :: Nil)),
+      expectedSchemaAfterLoad = Some(StructType(StructField("a", IntegerType) :: Nil)))
+  }
+
+  test("save with column names that are reserved words") {
+    testRoundtripSaveAndLoad(
+      s"save_with_column_names_that_are_reserved_words_$randomSuffix",
+      sqlContext.createDataFrame(sc.parallelize(Seq(Row(1))),
+        StructType(StructField("table", IntegerType) :: Nil)))
+  }
+
+  test("save with one empty partition (regression test for #96)") {
+    val df = sqlContext.createDataFrame(sc.parallelize(Seq(Row(1)), 2),
+      StructType(StructField("foo", IntegerType) :: Nil))
+    assert(df.rdd.glom.collect() === Array(Array.empty[Row], Array(Row(1))))
+    testRoundtripSaveAndLoad(s"save_with_one_empty_partition_$randomSuffix", df)
+  }
+
+  test("save with all empty partitions (regression test for #96)") {
+    val df = sqlContext.createDataFrame(sc.parallelize(Seq.empty[Row], 2),
+      StructType(StructField("foo", IntegerType) :: Nil))
+    assert(df.rdd.glom.collect() === Array(Array.empty[Row], Array.empty[Row]))
+    testRoundtripSaveAndLoad(s"save_with_all_empty_partitions_$randomSuffix", df)
+    // Now try overwriting that table. Although the new table is empty, it should still overwrite
+    // the existing table.
+    val df2 = df.withColumnRenamed("foo", "bar")
+    testRoundtripSaveAndLoad(
+      s"save_with_all_empty_partitions_$randomSuffix", df2, saveMode = SaveMode.Overwrite)
+  }
+
+  test("informative error message when saving a table with string that is longer than max length") {
+    val tableName = s"error_message_when_string_too_long_$randomSuffix"
+    try {
+      val df = sqlContext.createDataFrame(sc.parallelize(Seq(Row("a" * 512))),
+        StructType(StructField("A", StringType) :: Nil))
+      val e = intercept[SQLException] {
+        write(df)
+          .option("dbtable", tableName)
+          .mode(SaveMode.ErrorIfExists)
+          .save()
+      }
+      assert(e.getMessage.contains("while loading data into Redshift"))
+    } finally {
+      conn.prepareStatement(s"drop table if exists $tableName").executeUpdate()
+      conn.commit()
+    }
+  }
+
+  test("full timestamp precision is preserved in loads (regression test for #214)") {
+    val timestamps = Seq(
+      TestUtils.toTimestamp(1970, 0, 1, 0, 0, 0, millis = 1),
+      TestUtils.toTimestamp(1970, 0, 1, 0, 0, 0, millis = 10),
+      TestUtils.toTimestamp(1970, 0, 1, 0, 0, 0, millis = 100),
+      TestUtils.toTimestamp(1970, 0, 1, 0, 0, 0, millis = 1000))
+    testRoundtripSaveAndLoad(
+      s"full_timestamp_precision_is_preserved$randomSuffix",
+      sqlContext.createDataFrame(sc.parallelize(timestamps.map(Row(_))),
+        StructType(StructField("ts", TimestampType) :: Nil))
+    )
+  }
+}
